Update calibrate_nonlin.py
These are the lines in calibrate_nonlin.py that always check for the conditions of the two ut removed.

The issue comes from the fact that we now import parameters from detector.py and not an external yaml file. It would be odd to change the input arguments of calibrate_nonlin.py to accept a flag or new parameter values for only two ut that are anyways checked all the time in the main function.

check.real_nonnegative_scalar(rms_low_limit, 'rms_low_limit', TypeError)
    check.real_nonnegative_scalar(rms_upp_limit, 'rms_upp_limit', TypeError)
    if rms_low_limit >= rms_upp_limit:
        raise CalNonlinException('rms_upp_limit must be greater than rms_low_limit')