add arugments
Can still use the edit workflow, but add arguments because scripts that may want to run this script are not good at editing them first. Users can also write scripts that call this with proper arguemnts so that they do not accidentally commit their changes over these. Normal suggestion would be to remove the edit variables and make the arguments required.